Move genewiz import script to dbmigration app
Since it relies on the legacy database, makes sense to have
in the dbmigration app for now.

Eventually, when building a version of this that does not
rely on the legacy database (i.e. after flipping the switch to
retire the legacy system), it should be in the library app.