Change TemperatureChoiceField to a DecimalField.
Django 1.9 did not like how I was querying the ExperimentPlate table
here, when the ExperimentPlate table did not exist. While clearly that
would be a problem when creating a new instance of this form,
I don't understand why that was happening when I tried to just run
./manage.py migrate to populate the schema an empty database.

In Djangos 1.8.9 and before, this was not an issue. It starts with
Django 1.9.

Until I figure this out, I'll just use a DecimalField instead.