Do a bunch of work towards processing POST request.
Restructure view to process POST first. Change SupScore field
to TypedChoiceField, in order to coerce to a ManualScoreCode, after many
headaches trying to do this with to_python on a ChoiceField (unfortunately
to_python gets called before validate, and validate requires that the
choice be in choices, which are just pks). Lotsa other reorganization.
Including, change all old forms to do processing in process() instead of
clean() function.